requirements.txt
boto3==1.35.13
pytest==8.3.2
requests==2.32.3
moto==5.0.13
apache-airflow-providers-amazon
apache-airflow-providers-cncf-kubernetes
python-dotenv==1.0.1
